Add app configuration
- Allow app to specify a custom base class to handle authentication.
  This is needed for rails 7.1 compatibility
- Add support for custom hooks to restrict pinning
