+++ app/controllers/atomic_admin/atomic_tenant_client_id_strategy_controller.rb
@@ -1,5 +1,10 @@
 module AtomicAdmin
   class AtomicTenantClientIdStrategyController < ApplicationController
+
+    if AtomicAdmin.client_id_strategy_before_action.present?
+      before_action AtomicAdmin.client_id_strategy_before_action, only: [:create, :update]
+    end
+
     def pinned_client_id_params
       params.permit(:iss, :client_id, :application_instance_id)
     end
@@ -8,7 +13,7 @@ module AtomicAdmin
       AtomicTenant::PinnedClientId.find_by(id: params[:id])
     end
 
-    def search 
+    def search
       page = AtomicTenant::PinnedClientId
         .where(application_instance_id: params[:application_instance_id])
         .order(:id).paginate(page: params[:page], per_page: 30)
