+++ lib/atomic_admin/version.rb
@@ -1,3 +1,3 @@
 module AtomicAdmin
-  VERSION = "1.0.0"
+  VERSION = "1.1.0"
 end
